# admin_openapi.yaml
openapi: 3.0.3
info:
  title: Admin BFF
  version: 1.0.0
  description: |
    API gateway for players, wallets, withdrawals, and reports services.
    Conventions:
      - Array fields are always arrays (empty array when no data).
      - Optional fields may be null when data is unavailable.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /finances/api/v1/wallets/by-player-id/{playerId}:
    get:
      tags: [Wallets]
      summary: Get wallets by player ID
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Wallets
          content:
            application/json:
              schema:
                type: object
              examples:
                empty:
                  summary: No wallets
                  value:
                    wallets: []
        "404":
          $ref: "#/components/responses/NotFound"
  /finances/api/v1/wallets:
    post:
      tags: [Wallets]
      summary: Create wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWalletRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/wallets/deposit:
    post:
      tags: [Wallets]
      summary: Deposit to wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletDepositRequest"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/wallets/cashout:
    post:
      tags: [Wallets]
      summary: Cashout from wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletCashoutRequest"
      responses:
        "200":
          description: Cashout created
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/wallets/unlock:
    post:
      tags: [Wallets]
      summary: Unlock wallet funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletUnlockRequest"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/wallets/confirm-withdrawal:
    post:
      tags: [Wallets]
      summary: Confirm withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmWithdrawalRequest"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    CreateWalletRequest:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        currency:
          type: string
        type:
          type: string
          nullable: true
      required: [playerId, currency]
    WalletDepositRequest:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        amount:
          type: string
      required: [playerId, walletId, amount]
    WalletCashoutRequest:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        amount:
          type: string
        paymentSystemId:
          type: string
      required: [playerId, walletId, amount, paymentSystemId]
    WalletUnlockRequest:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        amount:
          type: string
      required: [playerId, walletId, amount]
    ConfirmWithdrawalRequest:
      type: object
      properties:
        withdrawalRequestId:
          type: string
          format: uuid
      required: [withdrawalRequestId]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string